# pkg/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: CMDB API
  version: 0.1.0
  description: 纯自研 CMDB 核心 API 契约，契约先行，前后端据此并行开发
servers:
  - url: /
    description: 相对路径（与后端同域部署）
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/v1/users:
    get:
      tags: [users]
      summary: 查询用户列表
      operationId: listUsers
      parameters:
        - name: keyword
          in: query
          description: 按用户名或显示名模糊过滤
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 用户分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [users]
      summary: 创建用户
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '200':
          description: 创建成功,返回完整用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/users/{user_id}:
    get:
      tags: [users]
      summary: 查询用户详情
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: 用户详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [users]
      summary: 更新用户(显示名/状态/重置密码/角色分配)
      operationId: patchUser
      description: 内置 admin 用户(is_builtin=true)不允许停用、不允许移除 admin 角色
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPatchRequest'
      responses:
        '200':
          description: 更新成功,返回完整用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/search:
    get:
      tags: [cis]
      summary: 全局全文搜索
      operationId: globalSearch
      description: 跨模型、CI、IPAM 前缀/IP 的统一搜索入口（landing 页全局搜索框）。结果按资源类型分组，每组上限由 limit 控制；无对应权限点的分组自动省略（如用户无 ipam:read 则不返回 IPAM 组）
      parameters:
        - name: q
          in: query
          description: 搜索关键字（大小写不敏感，子串匹配）
          required: true
          schema:
            type: string
            minLength: 1
        - name: limit
          in: query
          description: 每组返回上限，默认 10、最大 50
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: 分组搜索结果（仅返回非空分组）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: meridian_token
      description: 登录成功后由服务端写入的 httpOnly JWT 会话 cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 采集器等非浏览器客户端使用的 Bearer JWT(登录接口响应体中的 token)
  parameters:
    UserId:
      name: user_id
      in: path
      description: 用户 ID
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      description: 页码，从 1 开始
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      description: 每页条数
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
  responses:
    BadRequest:
      description: 请求参数或校验规则不通过
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 未登录或会话已失效(错误码 UNAUTHORIZED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 已登录但缺少所需权限点(错误码 FORBIDDEN)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    User:
      type: object
      required: [id, username, display_name, status, roles, is_builtin, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type: string
        status:
          type: string
          enum: [active, disabled]
        roles:
          type: array
          items:
            type: string
          description: 角色编码数组
        is_builtin:
          type: boolean
          description: 是否内置账号(内置 admin 不可停用/降权)
        scope_app_ids:
          type: array
          items:
            type: string
            format: uuid
          description: 数据范围（F-005）：归属的 biz_app CI ID 数组，空数组表示不受限
        scope_app_names:
          type: array
          items:
            type: string
          description: 与 scope_app_ids 同序的应用名称
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserCreateRequest:
      type: object
      required: [username, display_name, password]
      properties:
        username:
          type: string
          minLength: 2
        display_name:
          type: string
        password:
          type: string
          format: password
          minLength: 6
        roles:
          type: array
          items:
            type: string
          description: 角色编码数组
    UserPatchRequest:
      type: object
      description: 全字段可选,仅更新传入字段
      properties:
        display_name:
          type: string
        status:
          type: string
          enum: [active, disabled]
        password:
          type: string
          format: password
          minLength: 6
          description: 传入即重置密码
        roles:
          type: array
          items:
            type: string
          description: 传入即整体替换角色分配
        scope_app_ids:
          type: array
          items:
            type: string
            format: uuid
          description: 数据范围（F-005）：传入即整体替换并即时生效；空数组表示不受限；元素须为存量 biz_app CI
    UserListResponse:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
...
    SearchItem:
      type: object
      required: [kind, id, title, subtitle]
      properties:
        kind:
          type: string
          enum: [model, ci, ipam_prefix, ipam_ip]
          description: 资源类型
        id:
          type: string
          description: 资源 ID（CI 为 CI ID）
        title:
          type: string
          description: 主显示名（如主机名、机柜编号、模型名）
        subtitle:
          type: string
          description: 辅助信息（如所属模型、编码、CIDR）
        matched:
          type: string
          description: 命中说明（如匹配到的属性 key=value），可能为空
        model_code:
          type: string
          description: 仅 kind=ci 时返回，所属模型编码（前端据此决定跳转目标）
    SearchGroup:
      type: object
      required: [kind, label, items]
      properties:
        kind:
          type: string
          enum: [models, cis, ipam]
          description: 分组标识
        label:
          type: string
          description: 分组显示名（如 模型 / CI 实例 / IPAM）
        items:
          type: array
          items:
            $ref: '#/components/schemas/SearchItem'
    SearchResponse:
      type: object
      required: [query, groups]
      properties:
        query:
          type: string
        groups:
          type: array
          items:
            $ref: '#/components/schemas/SearchGroup'
          description: 仅包含非空分组
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 机器可读错误码
          example: VALIDATION_FAILED
        message:
          type: string
          description: 人类可读错误信息
          example: 属性 hostname 为必填项
        details:
          type: object
          additionalProperties: true
          description: 逐字段校验错误等补充信息
...